requirements.txt
cffi==1.17.1
cryptography==44.0.2
pillow==11.2.1
pycparser==2.22
